optionally propagate beets package in beetsPackages.beetcamp
If beets is propagated while it is acting as a plugin, the parent
beets' files will not appear in the final package; if beets is not
propagated, beetcamp cli won't work.

This feels like an anti-pattern to me but I think it works? I'm not
sure if these packages will conflict in a system profile though.